backend/requirements.txt
Flask
pytesseract
requests
Pillow
spacy